# Remove cloudfront config