@larspetrus 2017-05-01
It would be useful to have a way to go directly to the server, for debugging and general curiosity.

Maybe that would most easily be a DNS pointing directly to the IP?