@jfly 2017-05-01
Ah, I neglected to mention that, thanks! Yes, ssh-ing to the server is still going to be a pretty valuable tool. The naked domain (worldcubeassociation.org) is still pointing at the IP address of our server. It is just the www.worldcubeassociation.org domain that I set the CNAME on. As always, if someone goes directly to worldcubeassociation.org, they will be redirected to www.worldcubeassociation.org. Someone clever with HTTP hostnames could use this IP to bypass cloudfront if they want to, though =)

![image](https://cloud.githubusercontent.com/assets/277474/25587418/d192077a-2e58-11e7-8706-6eedd91bd084.png)
![image](https://cloud.githubusercontent.com/assets/277474/25587388/b24823fe-2e58-11e7-9ae6-3f0c6babe995.png)