@larspetrus 2017-05-01
I... don't get it. How would someone that clever bypass cloudfront?