@jfly 2017-05-01
> How would someone clever bypass cloudfront?

Nothing has changed with our EC2 instance: it has a public IP address and responds to HTTP(S) requests from anyone, including, but not limited to, cloudfront.