@larspetrus 2017-05-01
You mean you can use the IP address directly?

My suggestion (now more clearly thought and spelled out) is that we map the IP address to something like raw.worldcubeassociation.org