@jfly 2017-05-01
> You mean you can use the IP address directly?

Yeah, the simplest way I know of would be to edit your `/etc/hosts` file so www.worldcubeassociation.org resolves to 34.208.140.116 instead of being a CNAME to d2rwqro7nzw43m.cloudfront.net.

> My suggestion (now more clearly thought and spelled out) is that we map the IP address to something like raw.worldcubeassociation.org

We can certainly add whatever subdomains we want, but it's not really clear to me what the point of doing so would be. I'm happy to continue this discussion, but I'm going to go ahead and merge this PR up.