@larspetrus 2017-05-02
The point would mostly be simple debugging. If something looks weird on the
site or it's down, it's good to easily be able to check if it's because of
Cloudfront.

I've never touched /etc/hosts and would be a little nervous about it, but I
assume you could just have a line in there that you comment/uncomment. That
still has some problems. Pages served from both places would have the same
URL, I would always worry about which state I left /etc/hosts in, and it's
more steps than ideal.

So I would probably use https://34.208.140.116  directly. And while that of
course works, the whole purpose of DNS is to let people deal with
meaningful names rather than 12 digit numbers, so why not use it that way?

On Mon, May 1, 2017 at 2:07 PM, Jeremy Fleischman <notifications@github.com>
wrote:

> You mean you can use the IP address directly?
>
> Yeah, the simplest way I know of would be to edit your /etc/hosts file so
> www.worldcubeassociation.org resolves to 34.208.140.116 instead of being
> a CNAME to d2rwqro7nzw43m.cloudfront.net.
>
> My suggestion (now more clearly thought and spelled out) is that we map
> the IP address to something like raw.worldcubeassociation.org
>
> We can certainly add whatever subdomains we want, but it's not really
> clear to me what the point of doing so would be. I'm happy to continue this
> discussion, but I'm going to go ahead and merge this PR up.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/thewca/worldcubeassociation.org/pull/1583#issuecomment-298433035>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAuGK7WCuR6AyNWOX67VKI1QXsiqOOMhks5r1kmcgaJpZM4NNNz7>
> .
>



-- 
Procrastinators - the Leaders of Tomorrow