@jfly 2017-05-04
Ahh, I see. Yes, having an easy way to bypass cloudfront would be nice. Unfortunately, actually implementing that is more than just a DNS setting: we'll need some application level work to get this working. The problem is that we're using HSTS with includeSubDomains, which mean that any domain like `direct.worldcubeassociation.org` is going to need a corresponding HTTPS certificate and nginx configuration. I'm not sure it's worth the hassle =(