@larspetrus 2017-05-04
Oh, that that's different then. I assumed this was a simple DNS fix you
could do in a minute and have time left for a one hand solve.

Yeah, it's not worth a lot of hassle. Or even a medium amount.

On Wed, May 3, 2017 at 8:20 PM, Jeremy Fleischman <notifications@github.com>
wrote:

> Ahh, I see. Yes, having an easy way to bypass cloudfront would be nice.
> Unfortunately, actually implementing that is more than just a DNS setting:
> we'll need some application level work to get this working. The problem is
> that we're using HSTS with includeSubDomains, which mean that any domain
> like direct.worldcubeassociation.org is going to need a corresponding
> HTTPS certificate and nginx configuration. I'm not sure it's worth the
> hassle =(
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/thewca/worldcubeassociation.org/pull/1583#issuecomment-299088747>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAuGK4bDSoLnJwpECyq8axLXvKmtuSspks5r2UPngaJpZM4NNNz7>
> .
>



-- 
Procrastinators - the Leaders of Tomorrow