@jfly 2017-05-04
Yeah, if this was easy, I'd be all for it. It's cool that you can get there via https://34.208.140.116/ directl, though! I actually wasn't expecting that to work. I'm surprised our nginx config doesn't try to redirect you to www.worldcubeassociation.org...